requirements.txt
pandas
fastapi
uvicorn
slowapi
python-dotenv
pydantic
lxml
gunicorn
openai